requirements.txt
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
pyro_ppl==1.9.1
pytest==8.3.4
scikit_learn==1.6.1
setuptools==72.1.0
torch==2.6.0
tqdm==4.67.1
